refactor: separate handlers into handlers, models, and utils
For each item and transaction handlers
