+++ item/handlers.go
@@ -5,51 +5,6 @@ import (
 	"jual-beli/utils"
 )
 
-type Item struct {
-	Name         string
-	CapitalPrice float64
-	SalePrice    float64
-	Category     string
-	TotalSold    int
-}
-
-type TabItem []Item
-
-var ItemList TabItem
-
-func IsItemExist() bool {
-	return len(ItemList) > 0
-}
-
-func GetItemIndexByName(name string) int {
-	for index, item := range ItemList {
-		if item.Name == name {
-			return index
-		}
-	}
-	return -1
-}
-
-func DeleteItemByIndex(index int) {
-	var newItemList TabItem
-
-	for i := 0; i < len(ItemList); i++ {
-		if i != index {
-			newItemList = append(newItemList, ItemList[i])
-		}
-	}
-
-	ItemList = newItemList
-}
-
-func ShowItemList() {
-	for index, item := range ItemList {
-		number := index + 1
-		fmt.Printf("%d. %s dengan harga modal %.f, seharga %.f ada di kategori %s\n", number, item.Name, item.CapitalPrice, item.SalePrice, item.Category)
-	}
-	fmt.Println("--------------------------")
-}
-
 func AddItemMenu() {
 	utils.ClearScreen()
 	utils.PrintBreadcrumb("Menu", "Barang", "Tambah")
