+++ item/models.go
@@ -1,0 +1,13 @@
+package item
+
+type Item struct {
+	Name         string
+	CapitalPrice float64
+	SalePrice    float64
+	Category     string
+	TotalSold    int
+}
+
+type TabItem []Item
+
+var ItemList TabItem
